requirements.txt
google-generativeai
python-dotenv
termcolor
PyGithub
gitpython
rich
pyyaml
sentence-transformers
